Add header and style dropdown
