--- examples/vanilla-js-demo/header.ts
@@ -1,2 +1,0 @@
-
-
